% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_norm.R
\name{change_norm}
\alias{change_norm}
\title{Creates an \code{\link{change_norm_type}} object.}
\usage{
change_norm(mean_before = 0, mean_after = 1, covar, size)
}
\arguments{
\item{mean_before}{Pre-change means. Defaults to 0.}

\item{mean_after}{Anticipated post-change means. Defaults to 1.}

\item{covar}{Covariance matrix of observations. A \code{size}-by-\code{size} symmetric positive-definite matrix.}

\item{size}{Dimensions of the problem.}
}
\description{
This function initializes an instance of \code{\link{change_norm_type}} class.
}
\examples{
ch <- change_norm(mean_before = 0.0, mean_after = 1.0)
}
\keyword{change,}
\keyword{gaussian}
\keyword{normal,}
