% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{mc_run_to_stop}
\alias{mc_run_to_stop}
\title{Runs a one-sided monte-carlo simulation.}
\usage{
mc_run_to_stop(change, parameter_actual, rules, monte_carlo_count = 1)
}
\arguments{
\item{change}{The change-point descriptor.}

\item{parameter_actual}{Actual post-change parameter.}

\item{rules}{Stopping times to run monte-carlo on.}

\item{monte_carlo_count}{Number of monte-carlo statistics to run in parallel. Defaults to 1.}
}
\description{
This function does something silly...
}
\examples{

}
\keyword{monte-carlo,}
\keyword{simulations}
