% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cusum.R
\name{stat_cusum}
\alias{stat_cusum}
\title{Creates an \code{\link{stat_cusum_type}} object.}
\usage{
stat_cusum(monte_carlo_count = 1)
}
\arguments{
\item{monte_carlo_count}{Number of monte-carlo statistics to run in parallel. Defaults to 1.}
}
\description{
This function initializes an instance of \code{\link{stat_cusum_type}} class.
}
\examples{
cs <- stat_cusum()
}
\keyword{cusum}
