% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_parallel_cusum.R
\name{stat_parallel_cusum}
\alias{stat_parallel_cusum}
\title{Creates an \code{\link{stat_parallel_cusum_type}} object.}
\usage{
stat_parallel_cusum(change, possibly_affected, weights,
  monte_carlo_count = 1)
}
\arguments{
\item{change}{Change-point descriptor object.}

\item{possibly_affected}{Indicates affected subset possibilities. Either list of 1-by-\code{size} logical vectors, or an \code{n}-by-\code{size} logical matrix.}

\item{weights}{Weights for the underlying statistics. A \code{count}-dimensional vector.}

\item{monte_carlo_count}{Number of monte-carlo statistics to run in parallel. Defaults to 1.}
}
\description{
This function initializes an instance of \code{\link{stat_parallel_cusum_type}} class.
}
\examples{
ch <- changept::change_norm(size = 4);
a1 <- c(TRUE, TRUE, FALSE, TRUE);
a2 <- c(FALSE, TRUE, FALSE, FALSE);
cs <- stat_parallel_cusum(ch, list(a1, a2));
}
\keyword{cusum,}
\keyword{parallel}
