% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_parallel_cusum.R
\docType{class}
\name{stat_parallel_cusum_type}
\alias{stat_parallel_cusum_type}
\title{CUSUM Statistic Class}
\format{A parallel CUSUM statistic \code{\link{R6Class}} generator object}
\usage{
stat_parallel_cusum_type
}
\arguments{
\item{change}{Change-point descriptor object.}

\item{possibly_affected}{Indicates affected subset possibilities. Either list of 1-by-\code{size} logical vectors, or an \code{n}-by-\code{size} logical matrix.}

\item{weights}{Weights for the underlying statistics. A \code{count}-dimensional vector.}

\item{monte_carlo_count}{Number of monte-carlo statistics to run in parallel. Defaults to 1.}

\item{at}{A \code{monte_carlo_count}-dimensional vector.

Observe a Log-likelihood Basis

Update the CUSUM statistic with a new instantaneous basis log-likelihood ratio.}

\item{loglikelihood_basis}{Instantaneous basis log-likelihood ratio. A \code{size}-by-\code{n} matrix, where \code{n = sum(at)}.}

\item{at}{Mask indicating which monte-carlo threads to update. A \code{monte_carlo_count}-dimensional logical vector. Defaults to repeated TRUE.}

\item{at}{A \code{monte_carlo_count}-dimensional vector.

Dimensions of the problem (number of sensors). Read-only.
Number of possibly affected subsets. Read-only.
Number of monte-carlo statistics to run in parallel. Read-only.
Mask indicating possibly affected subsets. A \code{count}-by-\code{size} logical matrix. Read-only.
Logarithms of proper weights for the underlying statistics. A \code{count}-dimensional vector. Read-only.
Log-likelihood ratio constructor. Read-only.
Dimensions of the problem (number of sensors).
Number of possibly affected subsets.
Number of monte-carlo statistics to run in parallel.
Mask indicating possibly affected subsets. A \code{count}-by-\code{size} logical matrix.
Logarithms of proper weights for the underlying statistics. A \code{count}-dimensional vector.
Log-likelihood ratio constructor.
Current values of the underlying statistics. A \code{count}-by-\code{monte_carlo_count} matrix.
Current value of the statistics. A \code{monte_carlo_count}-dimensional vector.}
}
\description{
This class describes a parallel (GLR) CUSUM statistic.
}
\examples{
ch <- changept::change_norm(size = 4);
a1 <- c(TRUE, TRUE, FALSE, TRUE);
a2 <- c(FALSE, TRUE, FALSE, FALSE);
cs <- stat_parallel_cusum_type$new(ch, list(a1, a2));
R6ChangeNorm Constructor


Resets the value of the statistics.


Current value of the statistics. A \\code{monte_carlo_count}-dimensional vector. Read-only.

}
\keyword{constructor}
\keyword{cusum,}
\keyword{log-likelihood,}
\keyword{loglikelihood}
\keyword{observe,}
\keyword{parallel}
\keyword{parallel,}
