% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping_time.R
\name{stopping_time}
\alias{stopping_time}
\title{Creates an \code{\link{stopping_time_type}} object.}
\usage{
stopping_time(statistic, thresholds)
}
\arguments{
\item{statistic}{The underlying statistic.}

\item{thresholds}{A collection of thresholds to test the statistic against.}
}
\description{
This function initializes an instance of \code{\link{stopping_time_type}} class.
}
\examples{

}
\keyword{one-sided}
\keyword{stopping}
\keyword{time,}
