% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping_time.R
\docType{class}
\name{stopping_time_type}
\alias{stopping_time_type}
\title{Stopping Time Class}
\format{A one-sided stopping time \code{\link{R6Class}} generator object}
\usage{
stopping_time_type
}
\arguments{
\item{statistic}{The underlying statistic.}

\item{thresholds}{A collection of thresholds to test the statistic against.}
}
\description{
This class describes a one-sided stopping time.
}
\examples{

stopping_time_type Constructor


Resets the statistic and something else.

Don't do anything if every thread has stopped.
\\code{mask} is a \\code{monte_carlo_count}-sized logical vector, with
the number of ones coinciding with the width of \\code{llr_basis} matrix.
Example:
LLR column index  1   2 3    4       5        678 9   10   
mask              1---1-1----1-------1--------111-1---1---
m_is_running      1---1-1------------1--------1---1-------
LLR basis is a matrix with 10 columns.
Out of these 10 we want to pick 6 columns: ##1, 2, 3, 5, 6, 9.
Number of thresholds. Read-only.
The underlying statistic. Read-only.
A collection of thresholds to test the statistic against. A \\code{monte_carlo_count}-dimensional logical vector. Read-only.
Mask indicating whether a given monte-carlo thread is still running (hasn't crossed all thresholds). A \\code{monte_carlo_count}-dimensional logical vector. Read-only.
Counts the number of monte-carlo threads that are still running (haven't crossed all thresholds). Read-only.
Run lengths for a given threshold and monte-carlo thread. A \\code{monte_carlo_count}-by-\\code{count} matrix. Read-only.
Local time.
Number of thresholds.
The underlying statistic.
A collection of thresholds to test the statistic against. A \\code{count}-dimensional vector.
Mask indicating whether a given monte-carlo thread is still running (hasn't crossed all thresholds). A \\code{monte_carlo_count}-dimensional logical vector.
Mask indicating whether a given threshold has been crossed by a given monte-carlo thread. A \\code{monte_carlo_count}-by-\\code{count} logical matrix.
Run lengths for a given threshold and monte-carlo thread. A \\code{monte_carlo_count}-by-\\code{count} matrix.
}
\keyword{constructor}
\keyword{one-sided}
\keyword{one-sided,}
\keyword{stopping}
\keyword{time,}
